requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
bs4==0.0.1
build==1.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
docutils==0.20.1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
h11==0.14.0
hatchling==1.22.4
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.2
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
netaddr==0.10.1
nh3==0.2.15
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pkginfo==1.10.0
pluggy==1.4.0
pluginbase==1.0.1
pycparser==2.21
Pygments==2.17.2
pyproject_hooks==1.0.0
PySocks==1.7.1
python-engineio==4.8.0
python-nmap==0.7.1
python-socketio==5.10.0
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
selenium==4.16.0
simple-websocket==1.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabulate==0.9.0
trio==0.23.2
trio-websocket==0.11.1
trove-classifiers==2024.3.25
twine==5.0.0
urllib3==2.1.0
validators==0.22.0
wafw00f==2.2.0
Werkzeug==3.0.1
workbook==1.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarl==1.9.4
zipp==3.18.1
